+++ src/components/barGroup.jsx
