+++ src/components/responsive/barGroup.jsx
