+++ src/routes.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { Route } from "react-router";
 import { BrowserRouter, Switch } from "react-router-dom";
 
-import Main from "./containers/views/main";
+import Main from "./views/main";
 
 const Routes = () => (
   <BrowserRouter>
