+++ src/views/main/components/admissions.jsx
