+++ src/views/main/components/gestes2.jsx
